requirements.txt
customtkinter
mido
keyboard
pynput
pillow
requests
python-rtmidi
markdown
tkinterweb[svg]
pygments
pyinstaller
emoji
